requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncpg==0.31.0
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.42.80
botocore==1.42.80
cachetools==6.2.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chatterbot-corpus==1.2.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.0
ddgs==9.11.4
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.1.0
faiss-cpu==1.13.2
fastapi==0.111.0
fastapi-cli==0.0.24
fastjsonschema==2.21.1
filelock==3.21.2
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.3.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.14.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.5
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.30.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.1.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.13.2
kiwisolver==1.4.8
langchain==1.2.13
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.12
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.1.3
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.5
langsmith==0.7.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.1
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.0.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.16.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.3.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
ollama==0.6.1
openai==1.46.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
pooch==1.8.2
primp==1.1.3
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.4.1
protobuf==6.33.5
psutil==6.1.0
psycopg2-binary==2.9.11
PubChemPy==1.0.4
pure_eval==0.2.3
pyarrow==23.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pypdf==6.7.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gitlab==8.2.0
python-jose==3.5.0
python-json-logger==3.2.0
python-multipart==0.0.22
python-pptx==1.0.2
pytweening==1.2.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.7.1
QtPy==2.4.3
rdkit==2025.3.5
referencing==0.35.1
regex==2026.1.15
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.2
rich-toolkit==0.19.7
rpds-py==0.22.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.0
Send2Trash==1.8.3
sentence-transformers==5.2.2
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.37.2
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.40.0
typer==0.23.1
typer-slim==0.23.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.2.3
uuid_utils==0.14.0
uvicorn==0.29.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==2.1.1
xgboost==3.0.4
XlsxWriter==3.2.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
